Regression/Gradient-Descent-and-Logistic-Regression/requirements.txt
pandas
numpy
sklearn
matplotlib
plotly
scipy.spatial